Move into GA Service the extraction of metrics
In the same way that the Items extraction is isolated in the services
layer, I am following the same pattern here. I have added a method
to return the metrics for a particular day in batches of 10_000.

The method should grow through time; we should add parameters there like
the metrics, so we don’t couple this class with the object that use them,
but as of today this is not a real need.
