+++ spec/services/google_analytics_service_spec.rb
@@ -51,4 +51,64 @@ RSpec.describe GoogleAnalyticsService do
       expect(response).to eq([])
     end
   end
+
+  describe "#find_in_batches" do
+    before do
+      allow(subject.client).to receive(:fetch_all) do
+        [
+          build_report_data(
+            build_report_row(dimensions: %w(/foo), metrics: %w(1 1))
+          ),
+          build_report_data(
+            build_report_row(dimensions: %w(/bar), metrics: %w(2 2))
+          ),
+          build_report_data(
+            build_report_row(dimensions: %w(/cool), metrics: %w(3 3))
+          ),
+        ]
+      end
+    end
+
+    context 'when called with a block' do
+      it 'should yield successive report data' do
+        arg1 = [
+          a_hash_including('ga:pagePath' => '/foo', 'ga:pageviews' => 1, 'ga:uniquePageviews' => 1,
+          ),
+          a_hash_including(
+            'ga:pagePath' => '/bar',
+            'ga:pageviews' => 2,
+            'ga:uniquePageviews' => 2,
+          )
+        ]
+        arg2 = [
+          a_hash_including(
+            'ga:pagePath' => '/cool',
+            'ga:pageviews' => 3,
+            'ga:uniquePageviews' => 3,
+          )
+        ]
+
+        expect { |probe| subject.find_in_batches(date: Date.today, batch_size: 2, &probe) }
+          .to yield_successive_args(arg1, arg2)
+      end
+    end
+
+    private
+
+    def build_report_data(*report_rows)
+      Google::Apis::AnalyticsreportingV4::ReportData.new(rows: report_rows)
+    end
+
+    def build_report_row(dimensions:, metrics:)
+      Google::Apis::AnalyticsreportingV4::ReportRow.new(
+        dimensions: dimensions,
+        metrics: metrics.map do |metric|
+          Google::Apis::AnalyticsreportingV4::DateRangeValues.new(
+            values: Array(metric)
+          )
+        end
+      )
+    end
+  end
 end
+
